+++ NEWS.md
@@ -12,6 +12,7 @@
 * Added `s_coxph_pairwise` function to perform pairwise testing which is used by `t_tte` and `t_coxph`
 * Added methods for `t_count_true` replacing `t_summary_true`
 * Updated `t_count_unique` to create analysis subsets, add `t_el_count_unique` for vectors
+* Fixed bug in `t_events_term_id` so that table sort order is by decreasing frequency instead of alphabetical.
 
 # tern 0.6.4